no overlap D solver
