+++ alphacsc/_d_solver.py
@@ -1,10 +1,10 @@
 import numpy as np
 
-from .init_dict import get_init_strategy
+from ._base_solver import BaseDSolver
+from ._no_overlap import NoOverlapSolver
 from .loss_and_gradient import gradient_uv, gradient_d
-from .utils.validation import check_random_state
 from .utils.convolution import numpy_convolve_uv
-from .utils.dictionary import NoWindow, UVWindower, SimpleWindower
+from .utils.dictionary import UVWindower, SimpleWindower
 from .utils.optim import fista, power_iteration
 from .update_d_multi import prox_uv, prox_d
 
@@ -52,15 +52,15 @@ def check_solver_and_constraints(rank1, solver_d, uv_constraint):
         elif uv_constraint == 'auto' and solver_d in ['joint', 'fista']:
             uv_constraint = 'joint'
     else:
-        assert solver_d in ['auto', 'fista'], (
-            f"solver_d should be auto or fista. Got solver_d='{solver_d}'."
+        if solver_d == 'auto':
+            solver_d = 'fista'
+        assert solver_d in ['fista', 'no-overlap'], (
+            f"solver_d should be auto, fista, or no-overlap. Got solver_d='{solver_d}'."
         )
-        assert solver_d in ['auto', 'fista'] and uv_constraint == 'auto', (
-            "If rank1 is False, uv_constraint should be 'auto' "
-            f"and solver_d should be auto or fista. Got solver_d='{solver_d}' "
-            f"and uv_constraint='{uv_constraint}'."
+        assert uv_constraint == 'auto', (
+            "If rank1 is False, uv_constraint should be 'auto'."
+            f"Got uv_constraint='{uv_constraint}'."
         )
-        solver_d = 'fista'
     return solver_d, uv_constraint
 
 
@@ -145,222 +145,14 @@ def get_solver_d(n_channels, n_atoms, n_times_atom,
                 D_init, resample_strategy, window, eps, max_iter, momentum,
                 random_state, verbose, debug
             )
+        elif solver_d == 'no-overlap':
+            return NoOverlapSolver(
+                n_channels, n_atoms, n_times_atom, solver_d, uv_constraint,
+                D_init, resample_strategy, window, eps, max_iter, momentum,
+                random_state, verbose, debug
+            )
         else:
             raise ValueError('Unknown solver_d: %s' % (solver_d, ))
-
-
-class BaseDSolver:
-    """Base class for a d solver."""
-
-    def __init__(self, n_channels, n_atoms, n_times_atom, solver_d,
-                 uv_constraint, D_init, resample_strategy, window, eps,
-                 max_iter, momentum, random_state, verbose, debug):
-
-        self.n_channels = n_channels
-        self.n_atoms = n_atoms
-        self.n_times_atom = n_times_atom
-        self.uv_constraint = uv_constraint
-        self.eps = eps
-        self.max_iter = max_iter
-        self.momentum = momentum
-        self.rng = check_random_state(random_state)
-        self.verbose = verbose
-        self.debug = debug
-        self.D_init = D_init
-
-        self.init_strategy = get_init_strategy(
-            n_times_atom, self.get_D_shape(), self.rng, D_init
-        )
-
-        if not window:
-            self._windower = NoWindow()
-        else:
-            self._init_windower()
-
-        self.resample_strategy = resample_strategy
-        assert self.resample_strategy in ['greedy', 'chunk', 'random'], (
-            "resample_strategy should be greedy, chunk, or random. "
-            f"Got resample_strategy='{self.resample_strategy}'."
-        )
-
-    def _get_objective(self, z_encoder):
-
-        def objective(D, full=False):
-
-            D = self._windower.window(D)
-
-            return z_encoder.compute_objective(D)
-
-        return objective
-
-    def _get_prox(self):
-
-        def prox(D, step_size=None):
-
-            D = self._windower.window(D)
-
-            D = self.prox(D)
-
-            return self._windower.remove_window(D)
-
-        return prox
-
-    def _get_grad(self, z_encoder):
-
-        def grad(D):
-
-            D = self._windower.window(D)
-
-            grad = self.grad(D, z_encoder)
-
-            return self._windower.window(grad)
-
-        return grad
-
-    def init_dictionary(self, X):
-        """Returns a dictionary for the signal X depending on D_init value.
-
-        Parameters
-        ----------
-        X: array, shape (n_trials, n_channels, n_times)
-            The data on which to perform CSC.
-
-        Return
-        ------
-        D : array shape (n_atoms, n_channels + n_times_atom) or
-                  shape (n_atoms, n_channels, n_times_atom)
-            The initial atoms to learn from the data.
-        """
-
-        D_hat = self.init_strategy.initialize(X)
-
-        if not isinstance(self.D_init, np.ndarray):
-            D_hat = self._windower.window(D_hat)
-
-        self.D_hat = self.prox(D_hat)
-
-        return self.D_hat
-
-    def get_max_error_subwindow(self, z_encoder):
-        """Get the maximal reconstruction error sub-window from the data
-        as a new atom.
-
-        This idea is used for instance in [Yellin2017]
-
-        Parameters
-        ----------
-        z_encoder : BaseZEncoder
-            ZEncoder object to be able to compute the largest error sub-window.
-
-        Return
-        ------
-        dk: array, shape (n_channels + n_times_atom,) or
-                         (n_channels, n_times_atom,)
-            New atom for the dictionary, chosen as the sub-windos of the signal
-            with the maximal reconstruction error.
-
-        [Yellin2017] BLOOD CELL DETECTION AND COUNTING IN HOLOGRAPHIC LENS-FREE
-        IMAGING BY CONVOLUTIONAL SPARSE DICTIONARY LEARNING AND CODING.
-        """
-        assert z_encoder.n_channels == self.n_channels
-
-        d0 = z_encoder.get_max_error_patch()
-
-        d0 = self._windower.window(d0)
-
-        return self.prox(d0)
-
-    def add_one_atom(self, z_encoder):
-        """Adds one atom to D_hat and updates D_hat in z_encoder.
-
-        Parameters
-        ----------
-        z_encoder: BaseZEncoder
-            ZEncoder object.
-
-        Returns
-        -------
-        D_hat : array, shape (k+1, n_channels + n_times_atom) or
-                             (k+1, n_channels, n_times_atom)
-            The atoms to learn from the data, where k < n_atoms is the initial
-        number of atoms in the dictionary before adding an atom.
-        """
-        assert self.D_hat.shape[0] < self.n_atoms
-
-        new_atom = self.get_max_error_subwindow(z_encoder)[0]
-
-        self.D_hat = np.concatenate([self.D_hat, new_atom[None]])
-
-        z_encoder.set_D(self.D_hat)
-        return self.D_hat
-
-    def resample_atom(self, k0, z_encoder):
-        """Resamples the atom at index k0 of D_hat and updates D_hat in
-        z_encoder.
-
-        Parameters
-        ----------
-        k0: int
-            index of the atom to resample
-        z_encoder: BaseZEncoder
-            ZEncoder object.
-
-        Returns
-        -------
-        D_hat : array, shape (n_atoms, n_channels + n_times_atom) or
-                             (n_atoms, n_channels, n_times_atom)
-            The atoms to learn from the data.
-        """
-        if self.resample_strategy == 'greedy':
-            self.D_hat[k0] = self.get_max_error_subwindow(z_encoder)[0]
-        elif self.resample_strategy in ['chunk', 'random']:
-            from .init_dict import init_dictionary as init_dict
-            self.D_hat[k0] = init_dict(
-                z_encoder.X, 1, self.n_times_atom, self.uv_constraint,
-                rank1=self.rank1, window=(self._windower != NoWindow()),
-                D_init=self.resample_strategy, random_state=None
-            )
-        else:
-            raise NotImplementedError(
-                f"Unknown resampling strategy '{self.resample_strategy}'"
-            )
-
-        z_encoder.set_D(self.D_hat)
-        return self.D_hat
-
-    def update_D(self, z_encoder):
-        """Learn d's in time domain and update D_hat in z_encoder.
-
-        Parameters
-        ----------
-        z_encoder: BaseZEncoder
-            ZEncoder object.
-
-        Returns
-        -------
-        D_hat : array, shape (n_atoms, n_channels + n_times_atom) or
-                             (n_atoms, n_channels, n_times_atom)
-            The atoms to learn from the data.
-        """
-
-        assert z_encoder.n_channels == self.n_channels
-
-        D_hat0 = self._windower.remove_window(self.D_hat)
-
-        D_hat, pobj = fista(
-            self._get_objective(z_encoder), self._get_grad(z_encoder),
-            self._get_prox(), None, D_hat0, self.max_iter,
-            momentum=self.momentum, eps=self.eps, adaptive_step_size=True,
-            name=self.name, debug=self.debug, verbose=self.verbose
-        )
-
-        self.D_hat = self._windower.window(D_hat)
-
-        z_encoder.set_D(self.D_hat)
-
-        if self.debug:
-            return self.D_hat, pobj
-        return self.D_hat
 
 
 class Rank1DSolver(BaseDSolver):
@@ -385,14 +177,6 @@ class Rank1DSolver(BaseDSolver):
     def prox(self, D_hat):
         return prox_uv(D_hat, uv_constraint=self.uv_constraint,
                        n_channels=self.n_channels)
-
-    def get_D_shape(self):
-        """Returns the expected shape of the dictionary.
-
-        Note: For the 'greedy' strategy this does not return the actual
-        dictionary shape, but the final expected shape.
-        """
-        return (self.n_atoms, self.n_channels + self.n_times_atom)
 
 
 class JointDSolver(Rank1DSolver):
@@ -655,14 +439,6 @@ class DSolver(BaseDSolver):
 
     def prox(self, D_hat):
         return prox_d(D_hat)
-
-    def get_D_shape(self):
-        """Returns the expected shape of the dictionary.
-
-        Note: For the 'greedy' strategy this does not return the actual
-        dictionary shape, but the final expected shape.
-        """
-        return (self.n_atoms, self.n_channels, self.n_times_atom)
 
     def grad(self, D, z_encoder):
         return gradient_d(
